Move usecases to proper package
