+++ app/src/test/java/com/google/android/ground/domain/usecases/survey/ActivateSurveyUseCaseTest.kt
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.domain.usecase
+package com.google.android.ground.domain.usecases.survey
 
 import com.google.android.ground.BaseHiltTest
 import com.google.android.ground.FakeData.SURVEY
-import com.google.android.ground.domain.usecases.survey.ActivateSurveyUseCase
 import com.google.android.ground.persistence.local.stores.LocalSurveyStore
 import com.google.android.ground.persistence.remote.FakeRemoteDataStore
 import com.google.android.ground.repository.SurveyRepository
