+++ app/src/test/java/com/google/android/ground/domain/usecases/survey/ListAvailableSurveysUseCaseTest.kt
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.domain.usecase
+package com.google.android.ground.domain.usecases.survey
 
 import com.google.android.ground.BaseHiltTest
-import com.google.android.ground.domain.usecases.survey.ListAvailableSurveysUseCase
 import com.google.android.ground.model.Survey
 import com.google.android.ground.model.toListItem
 import com.google.android.ground.persistence.local.stores.LocalSurveyStore
