+++ app/src/test/java/com/google/android/ground/domain/usecases/survey/ReactivateLastSurveyUseCaseTest.kt
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.domain.usecase
+package com.google.android.ground.domain.usecases.survey
 
 import com.google.android.ground.BaseHiltTest
-import com.google.android.ground.domain.usecases.survey.ActivateSurveyUseCase
-import com.google.android.ground.domain.usecases.survey.ReactivateLastSurveyUseCase
 import com.google.android.ground.model.Survey
 import com.google.android.ground.persistence.local.LocalValueStore
 import com.google.android.ground.persistence.local.stores.LocalSurveyStore
